move test methods to TestRunTables
