+++ tests/unit/neptune/new/client/abstract_tables_test.py
@@ -29,10 +29,7 @@ from neptune.envs import (
     API_TOKEN_ENV_NAME,
     PROJECT_ENV_NAME,
 )
-from neptune.exceptions import (
-    MetadataInconsistency,
-    NeptuneException,
-)
+from neptune.exceptions import MetadataInconsistency
 from neptune.internal.backends.api_model import (
     Attribute,
     AttributeType,
@@ -187,20 +184,3 @@ class AbstractTablesTestMixin:
             path=["file", "set"],
             destination="some_directory",
         )
-
-    @patch.object(NeptuneBackendMock, "search_leaderboard_entries")
-    def test_fetch_runs_table_is_case_insensitive(self, search_leaderboard_entries):
-        states = ["active", "inactive", "Active", "Inactive", "aCTive", "INacTiVe"]
-        for state in states:
-            with self.subTest(state):
-                try:
-                    self.get_table(state=state)
-                    assert search_leaderboard_entries.called
-                except Exception as e:
-                    self.fail(e)
-
-    def test_fetch_runs_table_raises_correct_exception_for_incorrect_states(self):
-        for incorrect_state in ["idle", "running", "some_arbitrary_state"]:
-            with self.subTest(incorrect_state):
-                with self.assertRaises(NeptuneException):
-                    self.get_table(state=incorrect_state)
